showquantity state moved to filters state
