+++ client/src/store/Filters/actions.js
@@ -4,3 +4,10 @@ export const setFilterParams = (filterObj) => {
     payload: { filterParams: filterObj },
   };
 };
+
+export const newQuantity = (quantity) => {
+  return {
+    type: "SET_QUANTITY_ON_PAGE",
+    payload: quantity,
+  };
+};
