+++ client/src/store/Filters/reducer.js
@@ -1,7 +1,18 @@
-const reducer = (state = {}, action) => {
+const initialState = {
+  filters: {},
+  showQuantity: 9,
+};
+
+const reducer = (state = initialState, action) => {
   switch (action.type) {
     case "SET_FILTER_PARAMS": {
-      return action.payload.filterParams;
+      return { ...state, filters: action.payload.filterParams };
+    }
+    case "SET_QUANTITY_ON_PAGE": {
+      return {
+        ...state,
+        showQuantity: action.payload,
+      };
     }
     default: {
       return state;
