--- client/src/store/Quantity/actions.js
@@ -1,6 +1,0 @@
-export const newQuantity = (quantity) => {
-  return {
-    type: "SET_QUANTITY_ON_PAGE",
-    payload: quantity,
-  };
-};
