+++ client/src/store/store.js
@@ -8,13 +8,11 @@ import thunk from "redux-thunk";
 import modalReducer from "./Modal/reducer";
 import filterReducer from "./Filters/reducer";
 import productsReducer from "./Products/reducer";
-import quantityReducer from "./Quantity/reducer";
 
 export const rootReducer = combineReducers({
   modal: modalReducer,
   filters: filterReducer,
   products: productsReducer,
-  showQuantity: quantityReducer,
 });
 
 const devTools = window.__REDUX_DEVTOOLS_EXTENSION__
